adapt TEs
